#21 making things *much* faster by enabling the 4-wire sd bus and 100MHz SD clocking

+++ src/drivers/sai-driver.h
@@ -28,7 +28,7 @@ class SAIDriver {
         SAIDriver(bool stereo = true, BitDepth bitDepth = BitDepth::BIT_DEPTH_32, SampleRate sampleRate = SampleRate::SAMPLE_RATE_48K);
         ~SAIDriver();
 
-        int txTransmit(uint8_t* pData, uint16_t Size, uint32_t Timeout);
+        int txTransmit(uint8_t* pData, uint32_t Size, uint32_t Timeout);
 
         // static bool dmaQueueFull;
